refactor task
