+++ src/task/AtomicTask.ts
@@ -1,0 +1,3 @@
+import { DFTask } from "./DFTask";
+
+export class AtomicTask extends DFTask {}
